Fixed bug caused by parsing float vars as integer

+++ app/src/main/java/com/example/debtrecords/MainActivity.java
@@ -211,7 +211,7 @@ public class MainActivity extends AppCompatActivity
     {
         //First: validate the input
         EditText amountChangeEditText = currentlyExpandedRecord.findViewById(R.id.amountChangeEditText);
-        if(amountChangeEditText.getText().toString().isEmpty() || Integer.parseInt(amountChangeEditText.getText().toString()) == 0)
+        if(amountChangeEditText.getText().toString().isEmpty() || Float.parseFloat(amountChangeEditText.getText().toString()) == 0)
         {
             emphasizeChangeAmountField(true);
             return;
